+++ jasypt-spring-boot-starter/src/test/java/com/ulisesbocchio/jasyptspringbootstarter/BootstrappingJasyptConfigurationTest.java
@@ -1,4 +1,4 @@
-package com.ulisesbocchio.jasyptspringboot;
+package com.ulisesbocchio.jasyptspringbootstarter;
 
 import com.ulisesbocchio.jasyptspringboot.configuration.EnableEncryptablePropertiesBeanFactoryPostProcessor;
 import org.junit.After;
