+++ jasypt-spring-boot-starter/src/test/java/com/ulisesbocchio/jasyptspringbootstarter/EnvironmentVariablesLoadTest.java
@@ -1,4 +1,4 @@
-package com.ulisesbocchio.jasyptspringboot;
+package com.ulisesbocchio.jasyptspringbootstarter;
 
 import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.junit.BeforeClass;
