+++ jasypt-spring-boot-starter/src/test/java/com/ulisesbocchio/jasyptspringbootstarter/resolver/DefaultPropertyResolverTest.java
@@ -1,4 +1,4 @@
-package com.ulisesbocchio.jasyptspringboot.resolver;
+package com.ulisesbocchio.jasyptspringbootstarter.resolver;
 
 import com.fasterxml.jackson.databind.JsonNode;
